refactor(web): PermissionsController to Java
